+++ include-utils-macro/src/lib.rs
@@ -2,49 +2,38 @@
 
 #![allow(missing_docs)]
 
-use std::{
-    fmt::Display,
-    path::{Path, PathBuf},
-};
+use std::path::{Path, PathBuf};
 
 use itertools::Itertools;
-use proc_macro::TokenStream;
-use proc_macro_error::{proc_macro_error, Diagnostic, Level, OptionExt, ResultExt};
+use manyhow::manyhow;
+use proc_macro2::TokenStream as TokenStream2;
 use quote::quote;
-use syn::{parse_macro_input, LitStr};
+use syn::LitStr;
 
 use crate::include_location::{IncludeLocation, IncludeRange};
 
 mod include_location;
 
+#[manyhow]
 #[proc_macro]
-#[proc_macro_error]
-pub fn include_str_part(input: TokenStream) -> TokenStream {
-    let file = {
-        let file: LitStr = parse_macro_input!(input);
-        file.value()
-    };
-
-    let location = IncludeLocation::parse(&file);
-    let file_content = read_file(location.path);
+pub fn include_str_part(file: LitStr) -> manyhow::Result<TokenStream2> {
+    let file = file.value();
+    let location = IncludeLocation::parse(&file)?;
+    let file_content = read_file(location.path)?;
 
     let processed_content = process_file(file_content, &location.range, |_content, _name| {
-        proc_macro_error::abort_call_site!("Anchors is not supported for the plain string");
-    });
+        Err(manyhow::error_message!("Anchors is not supported for the plain string").into())
+    })?;
 
-    quote!(#processed_content).into()
+    Ok(quote!(#processed_content))
 }
 
+#[manyhow]
 #[proc_macro]
-#[proc_macro_error]
-pub fn include_md(input: TokenStream) -> TokenStream {
-    let file = {
-        let file: LitStr = parse_macro_input!(input);
-        file.value()
-    };
-
-    let location = IncludeLocation::parse(&file);
-    let file_content = read_file(location.path);
+pub fn include_md(file: LitStr) -> manyhow::Result<TokenStream2> {
+    let file = file.value();
+    let location = IncludeLocation::parse(&file)?;
+    let file_content = read_file(location.path)?;
 
     // TODO Use markdown parser to analyze comments.
     let processed_content = process_file(file_content, &location.range, |content, anchor_name| {
@@ -66,34 +55,27 @@ pub fn include_md(input: TokenStream) -> TokenStream {
             })
             .join("\n");
 
-        if !has_anchor {
-            proc_macro_error::abort_call_site!(
-                "Included file doesn't contain anchor with name: {}",
-                anchor_name
-            );
-        }
+        manyhow::ensure!(
+            has_anchor,
+            "Included file doesn't contain anchor with name: {anchor_name}"
+        );
+        Ok(output)
+    })?;
 
-        output
-    });
-
-    quote!(#processed_content).into()
+    Ok(quote!(#processed_content))
 }
 
-fn err_to_diagnostic(message: impl Display) -> Diagnostic {
-    Diagnostic::new(Level::Error, message.to_string())
-}
-
-fn cargo_manifest_dir() -> PathBuf {
-    std::env::var("CARGO_MANIFEST_DIR")
-        .map_err(err_to_diagnostic)
-        .unwrap_or_abort()
-        .into()
+fn cargo_manifest_dir() -> manyhow::Result<PathBuf> {
+    let dir = std::env::var("CARGO_MANIFEST_DIR").map_err(|err| {
+        manyhow::error_message!("Unable to read `CARGO_MANIFEST_DIR` environment variable. {err}")
+    })?;
+    Ok(dir.into())
 }
 
 /// Search for file path relative cargo manifest directory and workspace root directory if `workspace`
 /// feature is enabled.
-fn search_file(file_path: &Path) -> Option<PathBuf> {
-    let manifest_dir = cargo_manifest_dir();
+fn search_file(file_path: &Path) -> manyhow::Result<PathBuf> {
+    let manifest_dir = cargo_manifest_dir()?;
 
     let search_paths = [
         manifest_dir.clone(),
@@ -103,8 +85,9 @@ fn search_file(file_path: &Path) -> Option<PathBuf> {
             let metadata = cargo_metadata::MetadataCommand::new()
                 .manifest_path(manifest_dir.join("Cargo.toml"))
                 .exec()
-                .map_err(err_to_diagnostic)
-                .unwrap_or_abort();
+                .map_err(|err| {
+                    manyhow::error_message!("Unable to execute cargo metadata command. {err}")
+                })?;
 
             metadata.workspace_root.into_std_path_buf()
         },
@@ -114,37 +97,34 @@ fn search_file(file_path: &Path) -> Option<PathBuf> {
         let full_path = search_path.join(file_path);
 
         if full_path.exists() {
-            let full_path = full_path
-                .canonicalize()
-                .map_err(err_to_diagnostic)
-                .unwrap_or_abort();
-            return Some(full_path);
+            return full_path.canonicalize().map_err(|err| {
+                manyhow::error_message!("Unable to canonicalize file path. {err}").into()
+            });
         }
     }
 
-    None
+    Err(manyhow::error_message!("File path `{}` not found", file_path.display()).into())
 }
 
-fn read_file(file_path: impl AsRef<Path>) -> String {
+fn read_file(file_path: impl AsRef<Path>) -> manyhow::Result<String> {
     let full_path = {
         let file_path = file_path.as_ref();
         if file_path.is_absolute() {
             file_path.to_owned()
         } else {
-            search_file(file_path).expect_or_abort("unable to find path")
+            search_file(file_path)?
         }
     };
 
     std::fs::read_to_string(full_path)
-        .map_err(err_to_diagnostic)
-        .unwrap_or_abort()
+        .map_err(|err| manyhow::error_message!("Unable to read file. {err}").into())
 }
 
-fn process_file<F: FnOnce(String, &str) -> String>(
+fn process_file<F: FnOnce(String, &str) -> manyhow::Result<String>>(
     content: String,
     range: &IncludeRange<'_>,
     anchor_processor: F,
-) -> String {
+) -> manyhow::Result<String> {
     let content = match range {
         // Just copy the entire file content.
         IncludeRange::Full => content,
@@ -163,7 +143,8 @@ fn process_file<F: FnOnce(String, &str) -> String>(
                 lines.join("\n")
             }
         }
-        IncludeRange::Anchor { name } => anchor_processor(content, name),
+        IncludeRange::Anchor { name } => anchor_processor(content, name)?,
     };
-    content.trim().to_owned()
+
+    Ok(content.trim().to_owned())
 }